Rabbit MQ Support added.
This Update has a testing issue when using "setup.py test". For this
purpose, please use "python -m pytest test" instead.

The issue prevents the shutting down of the container. This issue is
external to this project and only effects testing.

Added RabbitMQ support and cleanup. This module can be drastically
improved.