Style and Documentation Update
- Version Bump 0.1.2 -> 0.2.0
- Added Style Formatting Script
- Added Style Information to README
- Fixed Style Formatting Issues
- Updated Doc Strings
- Add new Errors
- Updated `image_remove`
- Add Automated Style Checks

__Add Style Formatting Script__
Introduced a new script which will run `yarf`, `pylint` and
`pycodestyle` to preformat and then inforce style requirements.

- `yarf` will make inplace changes to the code resolving majority of the
formatiing issues.
- `pylint` is mainly use for logic typing errors.
- `pycodestyle` provides a third layer of protection.

__Add Style Information to README__
In an attempt to provide contributors (if ever lol) a quick location to
find style information, style information was added to the README.

__Fixed Style Formatting Issues__
Addressed all issues presented by the `style.sh` script.

__Updated Doc Strings__
As part of a requirement for pylint, docstrings were required for all
modules. This meant writing documentation for most of the code. It was a
natural transition into updating all documentation.

__Add New Errors__

New Errors Added:
- DockerNotAvailable
This new error code is used when connection to the Docker runtime is not
available. The most notable use of this is on import of the driver
module.
- OptionalModuleNOtInstalled
This new error is used if optional modules do not have all the required
pacakges.

__Update `image_remove`__
Before, if the `_remove_image` flag was set, we would attempt to delete
the image used. This would fail if the image was already attached to a
different container prior to running. We now catch the error and ignore
it. The decision was made so that we do not have to worry about breaking
other images on the local machine by force deleteing the image.

__Added Automated Style Checks__
Add automated style workflow to github for automated style validation on
push/pr.

New tasks Identified:
- Add username and password support for RabbitMQ
- Different Service version support and testing.