Issue#5 Resolved
- Resolve Bug
- Update tests

**Resolve Bug**
A Docker driver issue caused by a `ValueError: file descriptor cannot be a
negative integer (-1)` was preventing the containers from de-constructing. It is
unclear what the exact cause of this issue is but an issue has been forwarded to
the Docker python repository.

The issue is directly related to the Docker driver. This means that recreating
the client object using the `from_env` function each time the client is used
resolved the issue. This is not the ideal fix.

**Update Tests**
Attempt to catch issues in deconstruction of the driver object and fail tests if
an issue occurs. This will help prevent issues in the deconstruction from
slipping through tests in the future.